refactor: 리렌더링 범위를 줄이기 위한 컴포넌트 분리
