+++ components/pages/recruits/RecruitContents/styled.ts
@@ -1,0 +1,16 @@
+import styled from "styled-components";
+import { breakPoints } from "@/styles/Media";
+
+export const Wrapper = styled.div`
+  margin: 0 auto;
+  padding: 1.5rem 1rem;
+  max-width: ${breakPoints.desktop}px;
+`;
+
+export const ContentsHeader = styled.div`
+  width: 100%;
+  padding: 1rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
