+++ components/pages/recruits/RecruitFilters/styled.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+export const Wrapper = styled.div`
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 1rem;
+`;
+
+export const SelectBoxWrapper = styled.div`
+  width: 150px;
+`;
+
+export const SearchInput = styled.input`
+  padding: 0.5rem;
+  width: 225px;
+`;
